Conflicts:
	source/blender/blenlib/BLI_math_geom.h

+++ source/blender/blenlib/intern/math_vector_inline.c
@@ -315,7 +315,7 @@ MINLINE float dot_v3v3(const float a[3], const float b[3])
 
 MINLINE float cross_v2v2(const float a[2], const float b[2])
 {
-	 return a[0]*b[1] - a[1]*b[0];
+	return a[0]*b[1] - a[1]*b[0];
 }
 
 MINLINE void cross_v3_v3v3(float r[3], const float a[3], const float b[3])
@@ -472,9 +472,9 @@ MINLINE int equals_v4v4(const float *v1, const float *v2)
 
 MINLINE int compare_v3v3(const float *v1, const float *v2, const float limit)
 {
-	if(fabs(v1[0]-v2[0])<limit)
-		if(fabs(v1[1]-v2[1])<limit)
-			if(fabs(v1[2]-v2[2])<limit)
+	if(fabsf(v1[0]-v2[0])<limit)
+		if(fabsf(v1[1]-v2[1])<limit)
+			if(fabsf(v1[2]-v2[2])<limit)
 				return 1;
 
 	return 0;
@@ -493,10 +493,10 @@ MINLINE int compare_len_v3v3(const float *v1, const float *v2, const float limit
 
 MINLINE int compare_v4v4(const float *v1, const float *v2, const float limit)
 {
-	if(fabs(v1[0]-v2[0])<limit)
-		if(fabs(v1[1]-v2[1])<limit)
-			if(fabs(v1[2]-v2[2])<limit)
-				if(fabs(v1[3]-v2[3])<limit)
+	if(fabsf(v1[0]-v2[0])<limit)
+		if(fabsf(v1[1]-v2[1])<limit)
+			if(fabsf(v1[2]-v2[2])<limit)
+				if(fabsf(v1[3]-v2[3])<limit)
 					return 1;
 
 	return 0;
